RS/requirements.txt
dnspython==2.4.2
et-xmlfile==1.1.0
motor==3.3.2
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3